feat: add chromatic lib and refactor components
