+++ .env
@@ -1,0 +1,1 @@
+CHROMATIC_PROJECT_TOKEN="2515eeb2c0a6"
