+++ package.json
@@ -1,7 +1,9 @@
 {
   "name": "domeui-react",
   "version": "0.0.4",
-  "author": "DomeUI",
+  "author": {
+    "name": "DomeUI"
+  },
   "license": "MIT",
   "keywords": [
     "react",
@@ -26,7 +28,9 @@
     "test": "parcel demo/index.html",
     "dev": "start-storybook -p 6006",
     "build": "rimraf lib && rimraf .cache && rimraf dist-storybook && rollup -c && build-storybook -o dist-storybook",
-    "build:types": "tsc --emitDeclarationOnly --skipLibCheck"
+    "build:types": "tsc --emitDeclarationOnly --skipLibCheck",
+    "build-storybook": "build-storybook",
+    "chromatic": "chromatic"
   },
   "peerDependencies": {
     "react": "^17.0.2"
@@ -55,6 +59,7 @@
     "@typescript-eslint/parser": "5.17.0",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.4",
+    "chromatic": "^6.8.0",
     "core-js": "^3.0.1",
     "cross-env": "^7.0.3",
     "eslint": "8.12.0",
@@ -78,5 +83,7 @@
   "bugs": {
     "url": "https://github.com/Dome-UI/dome-ui-react/issues"
   },
-  "homepage": "https://github.com/Dome-UI/domeui-react#readme"
+  "homepage": "https://github.com/Dome-UI/domeui-react#readme",
+  "readme": "ERROR: No README data found!",
+  "_id": "domeui-react@0.0.4"
 }
