+++ src/components/button/index.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import { Button } from './Button'
+import { Button } from './index'
 
 <Meta
   title="Components/Button"
