+++ src/components/card/index.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from "@storybook/addon-docs";
-import { Card } from "./Card";
+import { Card } from "./index";
 
 <Meta
   title="Components/Card"
