+++ src/components/container/index.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from "@storybook/addon-docs";
-import { SectionContainer } from "./SectionContainer";
+import { SectionContainer } from "./index";
 
 <Meta
   title="Components/SectionContainer"
