--- src/components/container/index.ts
@@ -1,2 +1,0 @@
-
-export * from './SectionContainer'
