+++ src/components/flex/index.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from "@storybook/addon-docs";
-import { Flex } from "./Flex";
+import { Flex } from "./index";
 
 <Meta
   title="Components/Flex"
