--- src/components/flex/index.ts
@@ -1,1 +1,0 @@
-export * from './Flex'
