+++ src/components/input/index.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import { Input } from './Input'
+import { Input } from './index'
 
 <Meta
   title="Components/Input"
