+++ src/components/simpleCard/index.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from "@storybook/addon-docs";
-import { SimpleCard } from "./SimpleCard";
+import { SimpleCard } from "./index";
 
 <Meta
   title="Components/SimpleCard"
