--- src/components/simpleCard/index.ts
@@ -1,1 +1,0 @@
-export * from './SimpleCard'
