+++ src/components/text/index.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
-import { Text } from './Text'
+import { Text } from './index'
 
 <Meta
   title="Components/Text"
