+++ src/components/text/index.tsx
