+++ src/index.ts
@@ -1,6 +1,7 @@
-export * from './components/button'
-export * from './components/card'
-export * from './components/flex'
-export * from './components/input'
-export * from './components/simpleCard'
-export * from './components/text'
+export { Button } from './components/button'
+export { Card }  from './components/card'
+export { SectionContainer } from './components/container'
+export { Flex } from './components/flex'
+export { Input } from './components/input'
+export { SimpleCard } from './components/simpleCard'
+export { Text } from './components/text'
